# docs/architecture/api_spec.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Daily Stock Analysis API",
    "description": "Smart watchlist analysis API for A-shares, HK and US stocks\n\n## Modules\n- Stock analysis: trigger AI-powered analysis\n- History: query past analysis reports\n- Stock data: fetch market quotes\n\n## Authentication\nOptional admin authentication: when ADMIN_AUTH_ENABLED=true, all /api/v1/* routes except login, status, health checks and the OpenAPI docs require a valid admin session cookie; when disabled, no authentication is enforced.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8000",
      "description": "本地开发服务器"
    }
  ],
...
  "paths": {
...
    "/api/v1/decision-signals/latest/{stock_code}": {
      "get": {
        "tags": [
          "DecisionSignals"
        ],
        "summary": "Get latest active decision signals for a stock",
        "description": "Return the latest active signals for the given stock; lazy expiry runs before reading.",
        "operationId": "getLatestDecisionSignals",
        "security": [
          {
            "AdminSessionCookie": []
          }
        ],
        "parameters": [
          {
            "name": "stock_code",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Stock Code"
            }
          },
          {
            "name": "market",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Optional market filter: cn/hk/us/jp/kr/tw",
              "title": "Market"
            },
            "description": "Optional market filter: cn/hk/us/jp/kr/tw"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 1,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DecisionSignalListResponse"
                }
              }
            }
          },
          "401": {
            "description": "Not logged in or admin session invalid (when ADMIN_AUTH_ENABLED=true)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request parameters",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Path or query parameter validation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Query failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/decision-signals/{signal_id}": {
      "get": {
        "tags": [
          "DecisionSignals"
        ],
        "summary": "Get a decision signal",
        "description": "Get a single DecisionSignal by ID; lazy expiry runs before reading.",
        "operationId": "getDecisionSignal",
        "security": [
          {
            "AdminSessionCookie": []
          }
        ],
        "parameters": [
          {
            "name": "signal_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Signal Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DecisionSignalItem"
                }
              }
            }
          },
          "401": {
            "description": "Not logged in or admin session invalid (when ADMIN_AUTH_ENABLED=true)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Signal not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Path parameter validation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Query failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/decision-signals/{signal_id}/status": {
      "patch": {
        "tags": [
          "DecisionSignals"
        ],
        "summary": "Update decision signal status",
        "description": "Updates only valid statuses and optional metadata; omitting metadata keeps the existing value, null clears it, and an object replaces it wholesale while preserving the formal decision_profile identity. Terminal statuses such as expired/invalidated/closed/archived cannot be PATCHed back to active.",
        "operationId": "updateDecisionSignalStatus",
        "security": [
          {
            "AdminSessionCookie": []
          }
        ],
        "parameters": [
          {
            "name": "signal_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Signal Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DecisionSignalStatusUpdateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DecisionSignalItem"
                }
              }
            }
          },
          "401": {
            "description": "Not logged in or admin session invalid (when ADMIN_AUTH_ENABLED=true)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid status",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Signal not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Request body or path parameter validation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Update failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/decision-signals/outcomes/run": {
      "post": {
        "tags": [
          "DecisionSignals"
        ],
        "summary": "Trigger decision signal outcome evaluation",
        "description": "Explicitly trigger signal-level outcome computation; by default skips completed and terminal unable, but recomputes recoverable unable cases such as missing quote data; force=true recomputes and overwrites the same signal_id+horizon+engine_version。",
        "operationId": "runDecisionSignalOutcomes",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DecisionSignalOutcomeRunRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DecisionSignalOutcomeRunResponse"
                }
              }
            }
          },
          "401": {
            "description": "Not logged in or admin session invalid (when ADMIN_AUTH_ENABLED=true)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request fields",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Signal not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Request body validation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Outcome computation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "AdminSessionCookie": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DecisionSignalListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/DecisionSignalItem"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "page": {
            "type": "integer",
            "title": "Page"
          },
          "page_size": {
            "type": "integer",
            "title": "Page Size"
          }
        },
        "type": "object",
        "required": [
          "total",
          "page",
          "page_size"
        ],
        "title": "DecisionSignalListResponse"
      },
      "ErrorResponse": {
        "properties": {
          "error": {
            "type": "string",
            "title": "Error",
            "description": "错误类型",
            "example": "validation_error"
          },
          "message": {
            "type": "string",
            "title": "Message",
            "description": "错误详情",
            "example": "请求参数错误"
          },
          "detail": {
            "anyOf": [
              {},
              {
                "type": "null"
              }
            ],
            "title": "Detail",
            "description": "附加错误信息"
          }
        },
        "type": "object",
        "required": [
          "error",
          "message"
        ],
        "title": "ErrorResponse",
        "description": "错误响应",
        "example": {
          "error": "not_found",
          "message": "资源不存在"
        }
      },
...
    },
    "securitySchemes": {
      "AdminSessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "dsa_session"
      }
    }
  }
}